Initial version of an AWS Lambda for downloading raw data.
This should be set up as a scheduled Lambda event, and will
connect to the WoW Community API to look for auction house
data for a specified realm. See `handler._get_configuration()`
for details about environment variables that need to be set
for proper execution.

It should NOT be necessary to set any AWS credentials;
instead the `boto3` library will correctly inherit the IAM
role specified when the Lambda is set up. The only permissions
needed beyond regular Lambda execution are read/write access
to the S3 bucket where the raw data gets stored.