Record summarized batch in batch index table.
This uses DynamoDB as a place to record all of the batches we
have summarized so that we can find them without having to
iterate all the objects in the relevant S3 bucket. This uses
a combination hash-and-range key, using the date as the hash
(partition) key and the actual datetime as the range key. This
lets us easily grab all the batches in a day with a simple query.