Build a Python egg/wheel file.
Since it looks like running ECS, Fargate, or AWS Batch requires
running in a VPC, and using AWS services from within a VPC
requires a NAT Gateway, and NAT Gateways are $32/month, the most
economical choice for this is actually probably to build an AMI
and spin that up to run the batch job on demand, then spin it
back down again.

This creates a Python package for the training script that we
can install into the AMI we will create.