Configure logging for main script.
This allows for a couple of environment variables to dictate
where logging goes from the script. If the `LOG_TO` variable
is set, then logging will be directed to the specified file.
Regular console logging is enabled whether `LOG_TO` is
specified or not.

In addition, logging level can be controlled by the `LOG_LEVEL`
variable, which can be set to, say, `"logging.WARN"` to select
one of the pre-existing Pythong `logging` module levels. Level
defaults to `INFO` if not specified.

Also:
* mark main script as executable
* advance version number in setup.py