Fix minimum unit price calculation in batch summarization.
Auctions where there is not a buyout price (i.e. highest bid only)
as sometimes represented in the raw data as having a buyout price
of zero. As a result, we were (incorrectly) calculating a per-unit
buyout price of zero, which was ending up as the minimum unit
price for this batch.